fix(UserServiceImpl) : 회원 탈퇴 시 이미지 없을 경우 이미지 삭제하는 버그 fix

+++ src/main/java/com/today/todayproject/domain/user/service/UserServiceImpl.java
@@ -184,8 +184,10 @@ public class UserServiceImpl implements UserService {
             throw new BaseException(BaseResponseStatus.WRONG_CURRENT_PASSWORD);
         }
         String deleteProfileImgUrl = loginUser.getProfileImgUrl();
-        s3UploadService.deleteOriginalFile(deleteProfileImgUrl);
-        userRepository.delete(loginUser);
+        if (deleteProfileImgUrl != null) {
+            s3UploadService.deleteOriginalFile(deleteProfileImgUrl);
+            userRepository.delete(loginUser);
+        }
     }
 
     public UserGetPagingDto searchUsers(Pageable pageable, UserSearchDto userSearchDto) {
